# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Mopro Shop API
  version: "1.0.0"
  description: |
    Mobile-first marketplace API for Mopro Shop. Buyer app + seller panel.

    ## Money format
    All monetary values are **integer minor units** (kuruş for TRY; coin-minor for TRY_COIN).
    Every monetary field is accompanied by a `currency` string field.
    Never use decimal values. Display layer divides by 100.

    ## Timestamps
    All timestamps are **RFC 3339 strings** (e.g. `"2026-01-15T10:00:00Z"`).
    Dates (no time component) use ISO 8601 `YYYY-MM-DD`.

    ## Locale resolution
    Product `title` and `description` are resolved server-side from the
    `Accept-Language` request header. The response always contains a single
    resolved string — never a `translations[]` array.
    Fallback chain: requested locale → `tr-TR` → first available locale.

    ## Step-up authentication flow
    Endpoints annotated with `stepUpAuth` require a short-lived step-up token
    (TTL 5 min, claim `scope=high_sensitivity`). Sending a regular access token
    to a `stepUpAuth` endpoint returns `403` with `code: step_up_required`.

    Client flow:
    1. Receive `403 step_up_required`.
    2. Present OTP entry screen.
    3. `POST /auth/step-up` with access token + fresh OTP code → `step_up_token`.
    4. Retry the original request with `Authorization: Bearer <step_up_token>`.

    ## X-Trace-Id
    Clients should generate a UUID per request and attach it as `X-Trace-Id`.
...

    ## Idempotency
...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /categories:
    get:
      operationId: ListCategories
      tags: [catalog]
      summary: List all 42 product categories (locale-resolved names)
      parameters:
        - $ref: '#/components/parameters/TraceId'
      responses:
        '200':
          description: Category list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /orders/checkout:
    post:
      operationId: Checkout
      tags: [orders]
      summary: Atomically reserve cart → create order → initiate PSP payment
      description: |
        Single-shot convenience endpoint for the mobile checkout flow.
        Internally calls cart.Reserve → order.Create → payment.Initiate.
        Returns immediately with redirect_url for 3DS if required,
        or order details for non-3DS payment methods.
      parameters:
        - $ref: '#/components/parameters/TraceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Checkout initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Out of stock or cart modified concurrently
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 access token. TTL 15 min.
        Claims: sub (user_id as string), iat, exp.
        Issued by POST /auth/otp/verify or POST /auth/token/refresh.
    stepUpAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived JWT with claim `scope=high_sensitivity`. TTL 5 min.
        Issued by POST /auth/step-up after re-verifying OTP.
        Required for: DELETE /me (account deletion).
        If a regular bearerAuth token is sent instead, the server returns
        403 with code=step_up_required.
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Admin / internal token. Only for POST /products, POST /orders,
        POST /orders/{id}/refund, and GET /categories/{id}/commission.
        Not exposed to mobile clients.
  parameters:
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      description: |
        UUIDv7 generated client-side. Server caches the response for 24 hours
        keyed on this value. Duplicate requests within that window return the
        cached response without re-executing the operation.
      schema:
        type: string
        format: uuid
        example: "01926b7f-1234-7abc-8def-0123456789ab"
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      description: |
        Client-generated trace identifier (UUID or opaque string).
        Echoed in error responses as `error.trace_id`.
        Falls back to a server-generated UUID if absent.
      schema:
        type: string
        example: "4f3a2b1c-e71a-4c3f-b99a-8c3f2a1b7d5e"
  responses:
...
    Unauthorized:
      description: Missing, invalid, or expired access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    UnprocessableEntity:
      description: Request body failed schema or business-rule validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    ServiceUnavailable:
      description: |
        System is in read-only mode (maintenance window, migration in progress,
        or circuit breaker open). Retry after the `Retry-After` header value.
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: service_unavailable
              message: "System is temporarily in read-only mode. Please try again shortly."
              trace_id: "4f3a2b1c"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    FieldError:
      type: object
      description: A field-level validation error within an ErrorEnvelope
      required: [name, message]
      properties:
        name:
          type: string
          example: phone
        message:
          type: string
          example: "must start with +905"
    ErrorEnvelope:
      type: object
      description: Standard error response wrapper used by all error status codes
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id]
          properties:
            code:
              type: string
              description: Machine-readable error slug
              example: validation_error
            message:
              type: string
              description: Human-readable error message (locale from Accept-Language)
              example: "Telefon numarası geçersiz."
            trace_id:
              type: string
              description: Request trace ID. Echoes X-Trace-Id or server-generated UUID.
              example: "4f3a2b1c-e71a-4c3f-b99a-8c3f2a1b7d5e"
            fields:
              type: array
              description: Per-field validation errors. Present only for 422 responses.
              items:
                $ref: '#/components/schemas/FieldError'
...
    Category:
      type: object
      required: [id, name, slug, commission_pct_bps]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          description: Locale-resolved category name
        slug:
          type: string
        parent_id:
          type: integer
          format: int64
          nullable: true
        icon_url:
          type: string
          format: uri
          nullable: true
        commission_pct_bps:
          type: integer
          description: Commission rate in basis points (e.g. 1500 = 15%)
...
    CheckoutRequest:
      type: object
      required: [address_id, cargo_option, payment_method]
      properties:
        address_id:
          type: integer
          format: int64
        cargo_option:
          type: string
          enum: [aras, yurtici, surat, mng, hepsijet, ptt]
        payment_method:
          type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [card, coin_balance]
            saved_card_id:
              type: string
              nullable: true
              description: PSP-tokenized card ID. Required when type=card and using a saved card.
    CheckoutResponse:
      type: object
      required: [order_id, status, total_minor, currency, payment]
      properties:
        order_id:
          type: integer
          format: int64
        status:
          type: string
          enum: [awaiting_payment, confirmed]
        total_minor:
          type: integer
          format: int64
        currency:
          type: string
        payment:
          type: object
          required: [requires_3ds]
          properties:
            requires_3ds:
              type: boolean
            redirect_url:
              type: string
              format: uri
              nullable: true
              description: PSP 3DS redirect URL. Null when requires_3ds=false.
...